jansarthi-core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jansarthi-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.1",
    "fastapi[standard]>=0.121.0",
    "minio>=7.2.18",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.20",
    "passlib[bcrypt]>=1.7.4",
    "sqlmodel>=0.0.27",
    "twilio>=9.8.6",
]

[tool.setuptools]
packages = ["app"]

[tool.pyright]
typeCheckingMode = "basic"
pythonVersion = "3.10"